Add LegacyToDomainConverter
